requirements.txt
python-dotenv
PyPDF2
google-generativeai
numpy
faiss-cpu
Flask
Flask-Cors